13알차/pyproject.toml
[project]
name = "django-restaurant-review"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.5",
    "django-cleanup>=9.0.0",
    "django-extensions>=4.1",
    "django-restframework>=0.0.1",
    "djangorestframework>=3.16.1",
    "ipython>=9.4.0",
    "mysqlclient>=2.2.7",
    "pillow>=11.3.0",
]
